Remove special case for child status in Puppet 4
Recent versions of Puppet changed from returning an output String to
returning a ProcessOutput object. The psql provider wraps the output
string in an array along with child status. The type was assuming that
for Puppet 4 it was going to receive the ProcessOutput object *instead*
of the array. However, since the ProcessOutput object comes back *as*
the output from Puppet, and it's still wrapped in an array, the output
is actually exactly the same as it always has been. This code was
outsmarting itself and failing to work on Puppet 4.